refactor: domains 폴더를 services로 이동 및 MovieList를 MovieListService로 이름 변경
- src/domains/ → src/services/ 폴더 구조 변경
- MovieList 클래스 및 파일명 → MovieListService로 변경
- 관련 import 경로 전체 업데이트
